+++ apple/Classes/SensitiveContentAnalysisPlugin.swift
@@ -23,8 +23,8 @@ public class SensitiveContentAnalysisPlugin: NSObject, FlutterPlugin {
     @available(iOS 17.0, macOS 14.0, *)
     private func analyzeImage(image: FlutterStandardTypedData, result: @escaping (Bool?, Error?) -> Void) {
         DispatchQueue.global(qos: .userInteractive).async {
-            do {
-                let analyzer = try SCSensitivityAnalyzer()
+            Task {
+                let analyzer = SCSensitivityAnalyzer()
                 let policy = analyzer.analysisPolicy
 
                 if policy == .disabled {
@@ -88,9 +88,37 @@ public class SensitiveContentAnalysisPlugin: NSObject, FlutterPlugin {
                     }
                     #endif
                 }
-            } catch let error {
-                DispatchQueue.main.async {
-                    result(nil, error)
+            }
+        }
+    }
+
+    @available(iOS 17.0, macOS 14.0, *)
+    private func analyzeVideo(at fileURL: URL, result: @escaping (Bool?, Error?) -> Void) {
+    DispatchQueue.global(qos: .userInteractive).async {
+            Task {
+                let analyzer = SCSensitivityAnalyzer()
+                let policy = analyzer.analysisPolicy
+
+                if policy == .disabled {
+                    DispatchQueue.main.async {
+                        result(nil, nil)
+                    }
+                } else {
+                    Task {
+                        do {
+                            let handler = analyzer.videoAnalysis(forFileAt: fileURL)
+                            let analysisResult = try await handler.hasSensitiveContent()
+                            let isSensitive = analysisResult.isSensitive
+
+                            DispatchQueue.main.async {
+                                result(isSensitive, nil)
+                            }
+                        } catch let error {
+                            DispatchQueue.main.async {
+                                result(nil, error)
+                            }
+                        }
+                    }
                 }
             }
         }
@@ -100,7 +128,7 @@ public class SensitiveContentAnalysisPlugin: NSObject, FlutterPlugin {
     private func analyzeNetworkImage(at fileURL: URL, result: @escaping (Bool?, Error?) -> Void) {
         DispatchQueue.global(qos: .userInteractive).async {
             do {
-                let analyzer = try SCSensitivityAnalyzer()
+                let analyzer = SCSensitivityAnalyzer()
                 let policy = analyzer.analysisPolicy
 
                 if policy == .disabled {
@@ -166,9 +194,9 @@ public class SensitiveContentAnalysisPlugin: NSObject, FlutterPlugin {
     @available(iOS 17.0, macOS 14.0, *)
     private func checkPolicy(result: @escaping (Int?, Error?) -> Void) {
         DispatchQueue.global(qos: .userInitiated).async {
-            do {
-                let analyzer = try SCSensitivityAnalyzer()
-                // Check the current analysis policy. 
+            Task {
+                let analyzer = SCSensitivityAnalyzer()
+                // Check the current analysis policy.
                 let policy = analyzer.analysisPolicy
                 DispatchQueue.main.async {
                     if policy == .disabled {
@@ -179,19 +207,33 @@ public class SensitiveContentAnalysisPlugin: NSObject, FlutterPlugin {
                         result(2, nil)
                     }
                 }
-            } catch {
-                DispatchQueue.main.async {
-                    result(nil, error)
-                }
             }
         }
     }
-
 
     public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
         if #available(iOS 17.0, macOS 14.0, *) {
             // Handle Flutter method calls
             switch call.method {
+
+            case "analyzeVideo":
+                guard let args = call.arguments else {
+                    return result(FlutterError(code: "-1", message: "Could not recognize flutter arguments in method: \(call.method)", details: nil))
+                }
+                if let myArgs = args as? [String: Any],
+                let urlString = myArgs["url"] as? String,
+                let url = URL(string: urlString) {
+                    analyzeVideo(at: url) { isSensitive, error in
+                        if let error = error {
+                            result(FlutterError(code: "ERROR", message: "\(error)", details: nil))
+                        } else {
+                            result(isSensitive)
+                        }
+                    }
+                } else {
+                    result(FlutterError(code: "-1", message: "Could not recognize flutter arguments in method: \(call.method)", details: nil))
+                }
+
             case "analyzeNetworkImage":
                 guard let args = call.arguments else {
                     return result(FlutterError(code: "-1", message: "Could not recognize flutter arguments in method: \(call.method)", details: nil))